backend/chatbot-backend/requirements.txt
langchain-core
starlette
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
requests>=2.31.0
pydantic>=2.0.0
langchain-openai>=0.2.0
langgraph>=0.2.0
chromadb>=0.5.0
pandas>=2.2.0
duckdb>=1.1.0
pyarrow>=17.0.0
geopandas>=0.14.0
shapely>=2.0.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
tabulate

